refactor: move choice and selection to their own files
