+++ src/genjax/_src/generative_functions/combinators/drop_arguments.py
@@ -20,8 +20,8 @@ The `DropArgumentsGenerativeFunction` exposes GFI methods which eliminate stored
 This is useful to avoid unnecessary allocations in e.g. `MapCombinator` which uses `jax.vmap` as part of its implementation, causing the arguments stored in its callee's trace to be expanded and stored (unnecessarily). `DropArgumentsGenerativeFunction` eliminates the stored arguments in the callee's trace -- and allows us to retain a single copy of the arguments in the `MapCombinator` caller's `MapTrace`.
 """
 
+from genjax._src.core.datatypes.choice import ChoiceMap
 from genjax._src.core.datatypes.generative import (
-    ChoiceMap,
     GenerativeFunction,
     JAXGenerativeFunction,
     Trace,
