+++ src/genjax/_src/generative_functions/static/static_transforms.py
@@ -20,10 +20,12 @@ import jax
 import jax.numpy as jnp
 import jax.tree_util as jtu
 
-from genjax._src.core.datatypes.generative import (
+from genjax._src.core.datatypes.choice import (
     Choice,
     ChoiceMap,
     EmptyChoice,
+)
+from genjax._src.core.datatypes.generative import (
     GenerativeFunction,
     Trace,
 )
