+++ src/genjax/_src/inference/translator.py
@@ -20,9 +20,8 @@ import jax.tree_util as jtu
 from jax.experimental.checkify import check
 
 from genjax._src.checkify import optional_check
+from genjax._src.core.datatypes.choice import Choice, ChoiceMap
 from genjax._src.core.datatypes.generative import (
-    Choice,
-    ChoiceMap,
     GenerativeFunction,
     Trace,
 )
